move reactions to file
